refactor(nvim): put general settings in globals
