--- dotfiles/nvim/lua/ik1614/colorscheme.lua
@@ -1,1 +1,0 @@
-vim.cmd('colorscheme ' .. O.colorscheme)
