+++ dotfiles/nvim/lua/ik1614/config.lua
@@ -1,6 +1,5 @@
 O = {
     auto_complete = true,
-    colorscheme = 'mdracula',
     python = {
         linter = '',
         formatter = '',
