+++ dotfiles/nvim/lua/ik1614/globals/colorscheme.lua
@@ -1,0 +1,1 @@
+vim.cmd('colorscheme mdracula')
