+++ dotfiles/nvim/lua/ik1614/init.lua
@@ -1,14 +1,17 @@
 -- Setup globals that I expect to be always available.
 require('ik1614.globals')
-require('ik1614.settings') -- TODO: Move to globals
-require("ik1614.mappings") -- TODO: Move to globals
+require('ik1614.globals.settings')
+require("ik1614.globals.mappings")
+require("ik1614.globals.colorscheme")
 
+-- Setup config variables for plugins
 require('ik1614.config')
-require("ik1614.colorscheme")
 
+-- Hookup plugins
 require("ik1614.plugins")
 
 
+-- Plugin configs
 require("ik1614.nvim-compe")
 require("ik1614.galaxyline")
 
@@ -17,6 +20,7 @@ require("ik1614.telescope.mappings")
 
 require("ik1614.vim-tmux-navigator")
 
+-- LSP
 require("ik1614.lsp")
 require("ik1614.lsp.typescript")
 -- require("ik1614.lsp.python")
